+++ base/sysimg.jl
@@ -116,6 +116,15 @@ using .MultiplicativeInverses
 include("abstractarraymath.jl")
 include("arraymath.jl")
 
+# define MIME"foo/bar" early so that we can overload 3-arg show
+immutable MIME{mime} end
+macro MIME_str(s)
+    :(MIME{$(Expr(:quote, Symbol(s)))})
+end
+
+include("char.jl")
+include("strings/string.jl")
+
 # SIMD loops
 include("simdloop.jl")
 importall .SimdLoop
@@ -142,8 +151,9 @@ typealias StridedMatrix{T,A<:Union{DenseArray,StridedReshapedArray},I<:Tuple{Var
 typealias StridedVecOrMat{T} Union{StridedVector{T}, StridedMatrix{T}}
 
 # For OS specific stuff
-include(String(vcat(length(Core.ARGS)>=2?Core.ARGS[2].data:"".data, "build_h.jl".data))) # include($BUILDROOT/base/build_h.jl)
-include(String(vcat(length(Core.ARGS)>=2?Core.ARGS[2].data:"".data, "version_git.jl".data))) # include($BUILDROOT/base/version_git.jl)
+include(string((length(Core.ARGS)>=2 ? Core.ARGS[2] : ""), "build_h.jl"))     # include($BUILDROOT/base/build_h.jl)
+include(string((length(Core.ARGS)>=2 ? Core.ARGS[2] : ""), "version_git.jl")) # include($BUILDROOT/base/version_git.jl)
+
 include("osutils.jl")
 include("c.jl")
 include("sysinfo.jl")
@@ -158,14 +168,7 @@ include("io.jl")
 include("iostream.jl")
 include("iobuffer.jl")
 
-# define MIME"foo/bar" early so that we can overload 3-arg show
-immutable MIME{mime} end
-macro MIME_str(s)
-    :(MIME{$(Expr(:quote, Symbol(s)))})
-end
-
 # strings & printing
-include("char.jl")
 include("intfuncs.jl")
 include("strings/strings.jl")
 include("parse.jl")